--- a/pkg/upstream/cluster/lb_register_test.go
+++ b/pkg/upstream/cluster/lb_register_test.go
@@ -110,27 +110,28 @@ func TestRegisterNewLB(t *testing.T) {
 	// subset is also valid
 	//  reuse subset test config
 	subsetInfo := NewLBSubsetInfo(exampleSubsetConfig())
-	sublb := newSubsetLoadBalancer(headerKey, hs, newClusterStats("test"), subsetInfo)
-	// choose host is valid
-	// 1. ctx contains subset matched config
-	// 2. ctx contains header with key "hostname"
-	// should choose e1 only
-	for i := 0; i < 100; i++ {
-		host := sublb.ChooseHost(ctx)
-		if host == nil || host.Hostname() != "e1" {
-			t.Fatal("choose host not expected, get: ", host)
+	for name, sublb := range newSubsetLoadBalancers(headerKey, hs, newClusterStats("test"), subsetInfo){
+		// choose host is valid
+		// 1. ctx contains subset matched config
+		// 2. ctx contains header with key "hostname"
+		// should choose e1 only
+		for i := 0; i < 100; i++ {
+			host := sublb.ChooseHost(ctx)
+			if host == nil || host.Hostname() != "e1" {
+				t.Fatalf("[%s] choose host not expected, get: %s", name, host)
+			}
 		}
-	}
-	// choose e1,e2,e5
-	for i := 0; i < 100; i++ {
-		host := sublb.ChooseHost(ctx2)
-		if host == nil {
-			t.Fatal("choose host failed")
-		}
-		switch host.Hostname() {
-		case "e1", "e2", "e5":
-		default:
-			t.Fatal("choose host not expected, get: ", host)
+		// choose e1,e2,e5
+		for i := 0; i < 100; i++ {
+			host := sublb.ChooseHost(ctx2)
+			if host == nil {
+				t.Fatalf("[%s] choose host failed", name)
+			}
+			switch host.Hostname() {
+			case "e1", "e2", "e5":
+			default:
+				t.Fatalf("[%s] choose host not expected, get: %s", name, host)
+			}
 		}
 	}
 }