### Issues associated with this PR

https://github.com/mosn/mosn/issues/2008

### Solutions

https://github.com/mosn/mosn/issues/2008#issuecomment-1097844372

### Benchmark

```go

func BenchmarkSubsetLoadBalancer(b *testing.B) {
	ps := createHostset(benchHostConfigs(8000, 3))
	subsetConfig := benchSubsetConfig()
	b.Run("subsetLoadBalancer", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			newSubsetLoadBalancer(types.RoundRobin, ps, newClusterStats("BenchmarkSubsetLoadBalancer"), NewLBSubsetInfo(subsetConfig))
		}
	})
	b.Run("subsetLoadBalancerPreIndex", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			newSubsetLoadBalancerPreIndex(types.RoundRobin, ps, newClusterStats("BenchmarkSubsetLoadBalancer"), NewLBSubsetInfo(subsetConfig))
		}
	})
}
```

```bash
 dzdx@B-QBDRMD6M-0201  ~/Documents/mosn-open/pkg/upstream/cluster   perf/subsetlb ±✚  go test -bench=^BenchmarkSubsetLoadBalancer . -run=^$ -benchmem 
2022-04-14 18:55:42,662 [INFO] [network] [ register pool factory] register protocol: mock factory
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/upstream/cluster
BenchmarkSubsetLoadBalancer/subsetLoadBalancer-12                      9         118816550 ns/op        10105291 B/op       7217 allocs/op
BenchmarkSubsetLoadBalancer/subsetLoadBalancerPreIndex-12            246           5452478 ns/op         2427298 B/op      11463 allocs/op
PASS
ok      mosn.io/mosn/pkg/upstream/cluster       4.993s

```


<img width="1579" alt="image" src="https://user-images.githubusercontent.com/7933207/163155137-c6b72ed7-6cf3-4043-b257-ec989d8f216b.png">

|   endpoints | cpu (before) | cpu (after) |  alloc (before) | alloc (after) |
|---|---|---|---|---|
| 20 | 0.23ms | 0.35ms | 65KB | 189KB | 
| 100 | 0.93ms  |  0.38ms | 154KB | 214KB |
|  500 | 8.1ms  | 0.56ms  | 632KB | 322KB |
|  2000 |  27ms | 1.3ms |2.4MB | 738KB |
|  8000 |  102ms | 5.1ms  | 10MB | 2.4MB |


### Code Style
+ Make sure `Goimports` has run
+ Show `Golint` result