@taoyuanyuan 2022-04-14
Maybe we should offer a new LB type that is easy for us to compare or rollback.