@dzdx 2022-04-14
> Maybe we should offer a new LB type that is easy for us to compare or rollback.

@taoyuanyuan   
design like: 
rollback by change the buildMode in subsetInfo
```

type LBSubsetInfoImpl struct {
	enabled        bool
	fallbackPolicy types.FallBackPolicy
+       buildMode      types.SubsetBuildMode
	defaultSubSet  types.SubsetMetadata
	subSetKeys     []types.SortedStringSetType // sorted subset selectors
}

```

```
var lb types.LoadBalancer
if info.lbSubsetInfo.IsEnabled() {
    if info.lbSubsetInfo.BuildMode() == BuildModeBitmap {
        lb = NewBitmapSubsetLoadBalancer(info, hostSet)
    }else {
        lb = NewSubsetLoadBalancer(info, hostSet)
    }
} else {
    lb = NewLoadBalancer(info, hostSet)
}
```

How about this ?