@nejisama 2022-05-10
建议不要使用配置，而是使用全局变量的方式来切换模式。然后这个变量可以考虑通过ENV（环境变量）或者FeatureGate等方式来进行切换。